pyproject.toml
[tool.poetry]
name = "factorybot"
version = "0.1.0"
description = ""
authors = ["Aybek18 <kasymaliev09@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.7"
drf-spectacular = "^0.26.5"
django-environ = "^0.11.2"
pytelegrambotapi = "^4.13.0"
cryptography = "^41.0.4"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
